Improve error handling of pipeMaxSize()
If getline() returned -1 because the file was empty, don’t report an
error – in that case, the errno is 0 and the error message looks
confusing. (This can happen, for instance, when the program runs under
systemd with InaccessiblePaths=/proc/sys/fs/pipe-max-size.)

Also, if any error occurs, still set the cached pipeMaxSize so that the
error isn’t printed again the next time the function is called.
(Currently, main() calls it twice, for both stdin and stdout.)