Add code support for multiple outputs
output is a pointer into outputs, and the core loop prints to that
output. At the end, the output is incremented (next outputs element),
circling back to the first output if outputs is exhausted (it’s
NULL-terminated).

Currently, there’s no code to initialize more than one output, so it
might seem like much ado about nothing, but I’ll add dgsh support soon,
when we’ll have an arbitrary number of outputs (if compiled wih dgsh
support, which will of course be optional).