Change handleOptions() to return exitAfterOptions
Instead of having a global variable, make that value the return code –
even when we add more options, that value will be “special”, so I think
that’s justifiable.

This also fixes a bug where `ja2l -xh` would exit with code 0, not 1,
because the -h option overwrote the exit code 1 of the (nonexisting) -x
option.