Add dgsh support
When built with CFLAGS=-DUSE_DGSH LDLIBS=-ldgsh, ja2l supports dgsh and
scatters across any number of outputs. This can be used to speed up
processing by piping into a multipipe block something like this:

    function process {
        jq '…'
    }

    ja2l FILE | {{
        process &
        process &
        process &
        process &
    }} | summarize

This configuration would be suitable for a four-processor system; the
number of processing units can be adjusted easily. (This assumes that
the jq operation is the expensive part of the entire pipeline, and the
entire pipeline is CPU-bound. Otherwise this isn’t very useful.)