Create correct directories in mkosi.build
I forgot to update this part in fe27680769 – if the build tries to
install into the /usr/local, that tree needs to be created first.

(Honestly… perhaps I’ll just ignore the GNU Make manual’s recommendation
and add `mkdir -p` to the Makefile. This program already requires glibc
and Linux, it’s not really intended for cross-platform use anyways.)