Add alternative names of output formats
There are at least three terms to refer to ja2l’s output format, two of
them even with a formal specification (in fact, the specifications are
almost identical – I don’t know what the history here is). Since none of
the terms seems to be a clear standard, let’s just mention all three of
them at the end of the introduction.

Supersedes #3.